Added NavBar, organized folders
