+++ frontend/src/App.css
@@ -1,6 +1,6 @@
 .App {
     text-align: center;
-    display: flex;
+    
     min-height: 100vh;
     align-items: center;
     justify-content: center;
@@ -8,6 +8,11 @@
 }
 
 .auth-form-container, .login-form, .register-form{
+    text-align: center;
+    min-height: 100vh;
+    align-items: center;
+    justify-content: center;
+
     display: flex;
     flex-direction: column;
 }
@@ -32,3 +37,4 @@ button {
 .link-btn {
     background-color: #E39E6D;
 }
+
