+++ frontend/src/components/NavBar-Components/NavBar.js
@@ -1,0 +1,31 @@
+import { useImperativeHandle } from "react"
+import { Link, useMatch, useResolvedPath } from "react-router-dom"
+
+export default function NavBar() {
+    const path = window.location.pathname
+    return (
+        <nav className="nav">
+            <Link to="/homepage" className="site-title">
+                Our Cooking Mama
+            </Link>
+
+            <ul>
+                <CustomLink to="/homepage">Home Page</CustomLink>
+                <CustomLink to="/profile">Profile</CustomLink>
+            </ul>
+        </nav>
+    )
+}
+
+function CustomLink({ to, children, ...props }) {
+    //const path = window.location.pathname
+
+    const resolvedPath = useResolvedPath(to);
+    const isActive = useMatch({ path: resolvedPath.pathname, end:true });
+
+    return (
+        <li className={isActive ? "active" : ""}>
+            <Link to={to} {... props}>{children}</Link>
+        </li>
+    )
+}
