+++ frontend/src/index.js
@@ -1,6 +1,7 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import App from "./App";
+import "./components/NavBar-Components/styles.css"
 //import { BrowserRouter } from "react-router-dom";
 
 ReactDOM.render(<App />, document.getElementById("root"));
