+++ frontend/src/pages/ProfilePage.jsx
@@ -1,0 +1,15 @@
+import React, { useState } from "react";
+
+import NavBar from '../components/NavBar-Components/NavBar';
+
+export const ProfilePage = (props) => {
+
+    return (
+        <div>
+            <NavBar />
+            <h1>Profile Page</h1>
+        </div>
+    )
+}
+
+export default ProfilePage;
